Only use assembly code on known-compatible OS+Arch combinations.
Assume by default that an operating system does not have an ABI compatible
with the PerlAsm sources. Add all the operating systems that we've
explicitly added support for to the allowlist. Avoid trying to build or
use the PerlAsm code for those targets.

On top of this, we can build fallback logic for using Rust (or C)
implementations for those targets that aren't compatible with the
assembly.

(cherry picked from commit f37b8a2f3eaea75eaa58a0e1874f08b8517fd4aa)